Update NotificationModel and related methods to use String for ID
Changed the `id` field in `NotificationModel` from `DateTime` to `String` to better represent unique identifiers. Updated the `fromJson` and `toJson` methods accordingly. Also, modified the `retrieveLastProblemId` method in `AddNewProblemRepository` to remove the `docName` parameter and update the path. Adjusted the `updateNotificationTimeStamp` and `addStudentNotification` methods in `CheckAnswersRemoteDataSourceImpl` and `CheckAnswersRepoImpl` to use `String` for `notificationId`. Updated the `NotificationsCubit` to generate a `String` ID for new notifications.